# Enable strict mode in configure_mlx.sh